refactored code, created single store for Zustand, resolved comments
