+++ RDSExpoApp/.gitignore
@@ -34,3 +34,6 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
+
+# binary
+install-state.gz
