+++ RDSExpoApp/package.json
@@ -25,6 +25,7 @@
     "expo-font": "~13.0.4",
     "expo-image": "~2.0.7",
     "expo-linking": "~7.0.5",
+    "expo-modules-autolinking": "^2.0.8",
     "expo-router": "~4.0.18",
     "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.1",
